refactor: separate chat actions menu overlay into two dialogs, and add dependency on actions menu interaction listener
